DAY 3/enterprise_asset_mgmt/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-dotenv
mysql-connector-python
passlib[bcrypt]
python-multipart
python-jose[cryptography]
python-multipart
pytest httpx